Move categories.json to data/
