+++ scripts/data/categories.json
